fix: Resolved issue with absent "Stage" fields preventing saving
